Close #36 ehnance data transferred by gRPC
To be able to map data from BHL to CoL we need to transfer
more data via gRPC. In this commit we synchronize terms available
in BHL metadata dump and table names in schema. In particular
titles are renamed to items. In the future we should be able to
incorporate BHL metadata into the bhlindex and make it easier to
generate references list for a variety of name-strings and data-sources.
NameString data is added, that allows to migrate found name-strings
by gRPC to a remote computer.